requirements.txt
Flask==1.1.1
gunicorn==19.9.0
numpy==1.18.2
scipy==1.4.1
scikit-learn>=0.18
pyspellchecker==0.5.4
requests==2.22.0
datetime==4.3